fix(addselect): remove addselect content sub component
