+++ packages/ibm-products/src/patterns/AddSelect/example/components/MultiAddSelectWithColumnActions/MultiAddSelectWithColumnActions.tsx
@@ -733,48 +733,47 @@ export const MultiAddSelectWithColumnActions = forwardRef<
                   onSearch={handleGlobalSearch}
                   path={searchTerm ? [] : breadcrumbPath}
                   onBreadcrumbClick={handleBreadcrumbClick}
+                  layout="horizontal"
                 >
-                  <AddSelect.Content layout="horizontal">
-                    {currentItems.length > 0 ? (
-                      <>
-                        {/* First column - root level */}
+                  {currentItems.length > 0 ? (
+                    <>
+                      {/* First column - root level */}
+                      <ControlledColumn
+                        items={currentItems}
+                        onShowInfo={handleShowInfo}
+                        columnSearchPlaceholder={columnSearchPlaceholder}
+                        columnTitle={columnTitle || itemsLabel}
+                        level={1}
+                        dataManager={dataManager}
+                        onNavigateToChild={handleNavigateToChild}
+                        selectedItems={selectedIds}
+                        sortBy={rootSortBy}
+                        filterBy={rootFilterBy}
+                      />
+
+                      {/* Additional columns based on navigation levels */}
+                      {navigationLevels.map((navLevel, index) => (
                         <ControlledColumn
-                          items={currentItems}
+                          key={`${navLevel.parentId}-${index}`}
+                          items={navLevel.items}
                           onShowInfo={handleShowInfo}
                           columnSearchPlaceholder={columnSearchPlaceholder}
-                          columnTitle={columnTitle || itemsLabel}
-                          level={1}
+                          columnTitle={navLevel.parentTitle}
+                          level={index + 2}
                           dataManager={dataManager}
                           onNavigateToChild={handleNavigateToChild}
                           selectedItems={selectedIds}
-                          sortBy={rootSortBy}
-                          filterBy={rootFilterBy}
+                          sortBy={navLevel.sortBy}
+                          filterBy={navLevel.filterBy}
                         />
-
-                        {/* Additional columns based on navigation levels */}
-                        {navigationLevels.map((navLevel, index) => (
-                          <ControlledColumn
-                            key={`${navLevel.parentId}-${index}`}
-                            items={navLevel.items}
-                            onShowInfo={handleShowInfo}
-                            columnSearchPlaceholder={columnSearchPlaceholder}
-                            columnTitle={navLevel.parentTitle}
-                            level={index + 2}
-                            dataManager={dataManager}
-                            onNavigateToChild={handleNavigateToChild}
-                            selectedItems={selectedIds}
-                            sortBy={navLevel.sortBy}
-                            filterBy={navLevel.filterBy}
-                          />
-                        ))}
-                      </>
-                    ) : (
-                      <div className={`${blockClass}__no-results`}>
-                        <h4>{noResultsTitle}</h4>
-                        <p>{noResultsDescription}</p>
-                      </div>
-                    )}
-                  </AddSelect.Content>
+                      ))}
+                    </>
+                  ) : (
+                    <div className={`${blockClass}__no-results`}>
+                      <h4>{noResultsTitle}</h4>
+                      <p>{noResultsDescription}</p>
+                    </div>
+                  )}
                 </AddSelect.Body>
               </Tearsheet.MainContent>
