+++ packages/ibm-products/src/patterns/AddSelect/example/components/MultiAddSelectWithGlobalActions/MultiAddSelectWithGlobalActions.tsx
@@ -721,44 +721,43 @@ export const MultiAddSelectWithGlobalActions = forwardRef<
                   onBreadcrumbClick={handleBreadcrumbClick}
                   actionsSlot={filterDropdown}
                   subHeaderActions={sortMenu}
+                  layout="horizontal"
                 >
-                  <AddSelect.Content layout="horizontal">
-                    {filteredAndSortedItems.length > 0 ? (
-                      <>
-                        {/* First column - root level */}
+                  {filteredAndSortedItems.length > 0 ? (
+                    <>
+                      {/* First column - root level */}
+                      <ControlledColumn
+                        items={filteredAndSortedItems}
+                        onShowInfo={handleShowInfo}
+                        columnSearchPlaceholder={columnSearchPlaceholder}
+                        columnTitle={columnTitle || itemsLabel}
+                        level={1}
+                        dataManager={dataManager}
+                        onNavigateToChild={handleNavigateToChild}
+                        selectedItems={selectedIds}
+                      />
+
+                      {/* Additional columns based on navigation levels */}
+                      {navigationLevels.map((navLevel, index) => (
                         <ControlledColumn
-                          items={filteredAndSortedItems}
+                          key={`${navLevel.parentId}-${index}`}
+                          items={navLevel.items}
                           onShowInfo={handleShowInfo}
                           columnSearchPlaceholder={columnSearchPlaceholder}
-                          columnTitle={columnTitle || itemsLabel}
-                          level={1}
+                          columnTitle={navLevel.parentTitle}
+                          level={index + 2}
                           dataManager={dataManager}
                           onNavigateToChild={handleNavigateToChild}
                           selectedItems={selectedIds}
                         />
-
-                        {/* Additional columns based on navigation levels */}
-                        {navigationLevels.map((navLevel, index) => (
-                          <ControlledColumn
-                            key={`${navLevel.parentId}-${index}`}
-                            items={navLevel.items}
-                            onShowInfo={handleShowInfo}
-                            columnSearchPlaceholder={columnSearchPlaceholder}
-                            columnTitle={navLevel.parentTitle}
-                            level={index + 2}
-                            dataManager={dataManager}
-                            onNavigateToChild={handleNavigateToChild}
-                            selectedItems={selectedIds}
-                          />
-                        ))}
-                      </>
-                    ) : (
-                      <div className={`${blockClass}__no-results`}>
-                        <h4>{noResultsTitle}</h4>
-                        <p>{noResultsDescription}</p>
-                      </div>
-                    )}
-                  </AddSelect.Content>
+                      ))}
+                    </>
+                  ) : (
+                    <div className={`${blockClass}__no-results`}>
+                      <h4>{noResultsTitle}</h4>
+                      <p>{noResultsDescription}</p>
+                    </div>
+                  )}
                 </AddSelect.Body>
               </Tearsheet.MainContent>
