+++ packages/ibm-products/src/patterns/AddSelect/example/components/MultiAddSelectWithHierarchy/MultiAddSelectWithHierarchy.tsx
@@ -730,48 +730,47 @@ export const MultiAddSelectWithHierarchy = forwardRef<
                   onSearch={handleGlobalSearch}
                   path={searchTerm ? [] : breadcrumbPath}
                   onBreadcrumbClick={handleBreadcrumbClick}
+                  layout="horizontal"
                 >
-                  <AddSelect.Content layout="horizontal">
-                    {currentItems.length > 0 ? (
-                      <>
-                        {/* First column - root level */}
+                  {currentItems.length > 0 ? (
+                    <>
+                      {/* First column - root level */}
+                      <ControlledColumn
+                        items={currentItems}
+                        onShowInfo={handleShowInfo}
+                        columnSearchPlaceholder={columnSearchPlaceholder}
+                        columnTitle={columnTitle || itemsLabel}
+                        level={1}
+                        dataManager={dataManager}
+                        onNavigate={handleNavigateToChild}
+                        selectedItems={selectedIds}
+                        onSelectAll={handleSelectAll}
+                        partiallySelectedItems={partiallySelectedItems}
+                      />
+
+                      {/* Additional columns based on navigation levels */}
+                      {navigationLevels.map((navLevel, index) => (
                         <ControlledColumn
-                          items={currentItems}
+                          key={`${navLevel.parentId}-${index}`}
+                          items={navLevel.items}
                           onShowInfo={handleShowInfo}
                           columnSearchPlaceholder={columnSearchPlaceholder}
-                          columnTitle={columnTitle || itemsLabel}
-                          level={1}
+                          columnTitle={navLevel.parentTitle}
+                          level={index + 2}
                           dataManager={dataManager}
                           onNavigate={handleNavigateToChild}
                           selectedItems={selectedIds}
                           onSelectAll={handleSelectAll}
                           partiallySelectedItems={partiallySelectedItems}
                         />
-
-                        {/* Additional columns based on navigation levels */}
-                        {navigationLevels.map((navLevel, index) => (
-                          <ControlledColumn
-                            key={`${navLevel.parentId}-${index}`}
-                            items={navLevel.items}
-                            onShowInfo={handleShowInfo}
-                            columnSearchPlaceholder={columnSearchPlaceholder}
-                            columnTitle={navLevel.parentTitle}
-                            level={index + 2}
-                            dataManager={dataManager}
-                            onNavigate={handleNavigateToChild}
-                            selectedItems={selectedIds}
-                            onSelectAll={handleSelectAll}
-                            partiallySelectedItems={partiallySelectedItems}
-                          />
-                        ))}
-                      </>
-                    ) : (
-                      <div className={`${blockClass}__no-results`}>
-                        <h4>{noResultsTitle}</h4>
-                        <p>{noResultsDescription}</p>
-                      </div>
-                    )}
-                  </AddSelect.Content>
+                      ))}
+                    </>
+                  ) : (
+                    <div className={`${blockClass}__no-results`}>
+                      <h4>{noResultsTitle}</h4>
+                      <p>{noResultsDescription}</p>
+                    </div>
+                  )}
                 </AddSelect.Body>
               </Tearsheet.MainContent>
