+++ packages/ibm-products/src/patterns/AddSelect/example/components/NonHierarchicalWithPeekInsideItem/NonHierarchicalWithPeekInsideItem.tsx
@@ -385,142 +385,139 @@ export const NonHierarchicalWithPeekInsideItem = forwardRef<
                   globalSearchPlaceholder={globalSearchPlaceholder}
                   searchResultsTitle={searchResultsTitle}
                   onSearch={handleSearch}
+                  layout="horizontal"
                 >
-                  <AddSelect.Content layout="horizontal">
+                  <AddSelect.Column
+                    className={`${blockClass}__items-column`}
+                    title={itemsLabel}
+                    itemCount={filteredItems.length}
+                    multi={true}
+                    hideSearch
+                    onNavigate={handleNavigate}
+                  >
+                    {filteredItems.length > 0 ? (
+                      filteredItems.map((item) => {
+                        const isSelected = selectedIds.has(item.id);
+                        const currentModifierValue =
+                          modifierStates[item.id] ||
+                          item[modifierConfig.id] ||
+                          [];
+
+                        return (
+                          <AddSelect.Row
+                            key={item.id}
+                            itemId={item.id}
+                            title={item.title || ''}
+                            subtitle={item.subtitle}
+                            value={item.value || ''}
+                            icon={item.icon}
+                            disabled={item.disabled}
+                            hasItemPanel={!!item.itemDetails}
+                            onItemPanelClick={handleShowInfo}
+                            hasChildren={true}
+                          >
+                            <div className={`${blockClass}__modifier`}>
+                              <Layer>
+                                {modifierConfig.multiSelect ? (
+                                  <MultiSelect
+                                    id={`modifier-${item.id}`}
+                                    titleText={modifierConfig.title}
+                                    type="inline"
+                                    label={modifierConfig.label}
+                                    items={modifierConfig.options}
+                                    initialSelectedItems={
+                                      Array.isArray(currentModifierValue)
+                                        ? currentModifierValue
+                                        : []
+                                    }
+                                    onChange={({ selectedItems }) =>
+                                      handleModifierChange(
+                                        item.id,
+                                        selectedItems || []
+                                      )
+                                    }
+                                    disabled={!isSelected}
+                                    size="sm"
+                                  />
+                                ) : (
+                                  <Dropdown
+                                    id={`modifier-${item.id}`}
+                                    titleText={modifierConfig.title}
+                                    type="inline"
+                                    label={modifierConfig.label}
+                                    items={modifierConfig.options}
+                                    initialSelectedItem={
+                                      typeof currentModifierValue === 'string'
+                                        ? currentModifierValue
+                                        : currentModifierValue[0]
+                                    }
+                                    onChange={({ selectedItem }) =>
+                                      handleModifierChange(
+                                        item.id,
+                                        selectedItem
+                                      )
+                                    }
+                                    disabled={!isSelected}
+                                    size="sm"
+                                  />
+                                )}
+                              </Layer>
+                            </div>
+                          </AddSelect.Row>
+                        );
+                      })
+                    ) : (
+                      <div className={`${blockClass}__no-results`}>
+                        <h4>{noResultsTitle}</h4>
+                        <p>{noResultsDescription}</p>
+                      </div>
+                    )}
+                  </AddSelect.Column>
+
+                  {peekInsideItem && peekInsideItem.users && (
                     <AddSelect.Column
-                      className={`${blockClass}__items-column`}
-                      title={itemsLabel}
-                      itemCount={filteredItems.length}
-                      multi={true}
+                      title={peekInsideItem.title || ''}
+                      itemCount={peekInsideItem.users.length}
                       hideSearch
-                      onNavigate={handleNavigate}
                     >
-                      {filteredItems.length > 0 ? (
-                        filteredItems.map((item) => {
-                          const isSelected = selectedIds.has(item.id);
-                          const currentModifierValue =
-                            modifierStates[item.id] ||
-                            item[modifierConfig.id] ||
-                            [];
-
-                          return (
-                            <AddSelect.Row
-                              key={item.id}
-                              itemId={item.id}
-                              title={item.title || ''}
-                              subtitle={item.subtitle}
-                              value={item.value || ''}
-                              icon={item.icon}
-                              disabled={item.disabled}
-                              hasItemPanel={!!item.itemDetails}
-                              onItemPanelClick={handleShowInfo}
-                              hasChildren={true}
-                            >
-                              <div className={`${blockClass}__modifier`}>
-                                <Layer>
-                                  {modifierConfig.multiSelect ? (
-                                    <MultiSelect
-                                      id={`modifier-${item.id}`}
-                                      titleText={modifierConfig.title}
-                                      type="inline"
-                                      label={modifierConfig.label}
-                                      items={modifierConfig.options}
-                                      initialSelectedItems={
-                                        Array.isArray(currentModifierValue)
-                                          ? currentModifierValue
-                                          : []
-                                      }
-                                      onChange={({ selectedItems }) =>
-                                        handleModifierChange(
-                                          item.id,
-                                          selectedItems || []
-                                        )
-                                      }
-                                      disabled={!isSelected}
-                                      size="sm"
-                                    />
-                                  ) : (
-                                    <Dropdown
-                                      id={`modifier-${item.id}`}
-                                      titleText={modifierConfig.title}
-                                      type="inline"
-                                      label={modifierConfig.label}
-                                      items={modifierConfig.options}
-                                      initialSelectedItem={
-                                        typeof currentModifierValue === 'string'
-                                          ? currentModifierValue
-                                          : currentModifierValue[0]
-                                      }
-                                      onChange={({ selectedItem }) =>
-                                        handleModifierChange(
-                                          item.id,
-                                          selectedItem
-                                        )
-                                      }
-                                      disabled={!isSelected}
-                                      size="sm"
-                                    />
-                                  )}
-                                </Layer>
-                              </div>
-                            </AddSelect.Row>
-                          );
-                        })
-                      ) : (
-                        <div className={`${blockClass}__no-results`}>
-                          <h4>{noResultsTitle}</h4>
-                          <p>{noResultsDescription}</p>
-                        </div>
-                      )}
-                    </AddSelect.Column>
-
-                    {peekInsideItem && peekInsideItem.users && (
-                      <AddSelect.Column
-                        title={peekInsideItem.title || ''}
-                        itemCount={peekInsideItem.users.length}
-                        hideSearch
-                      >
-                        <div className={`${blockClass}__peek-list`}>
-                          {peekInsideItem.users.map((user) => (
-                            <div
-                              key={user.id}
-                              className={`${blockClass}__peek-item`}
-                            >
+                      <div className={`${blockClass}__peek-list`}>
+                        {peekInsideItem.users.map((user) => (
+                          <div
+                            key={user.id}
+                            className={`${blockClass}__peek-item`}
+                          >
+                            <div className={`${blockClass}__peek-item-wrapper`}>
                               <div
-                                className={`${blockClass}__peek-item-wrapper`}
+                                className={`${blockClass}__peek-item-content`}
                               >
                                 <div
-                                  className={`${blockClass}__peek-item-content`}
+                                  className={`${blockClass}__peek-item-icon`}
                                 >
-                                  <div
-                                    className={`${blockClass}__peek-item-icon`}
+                                  {user.icon}
+                                </div>
+                                <div
+                                  className={`${blockClass}__peek-item-text`}
+                                >
+                                  <span
+                                    className={`${blockClass}__peek-item-title`}
                                   >
-                                    {user.icon}
-                                  </div>
-                                  <div
-                                    className={`${blockClass}__peek-item-text`}
-                                  >
+                                    {user.title}
+                                  </span>
+                                  {user.subtitle && (
                                     <span
-                                      className={`${blockClass}__peek-item-title`}
+                                      className={`${blockClass}__peek-item-subtitle`}
                                     >
-                                      {user.title}
+                                      {user.subtitle}
                                     </span>
-                                    {user.subtitle && (
-                                      <span
-                                        className={`${blockClass}__peek-item-subtitle`}
-                                      >
-                                        {user.subtitle}
-                                      </span>
-                                    )}
-                                  </div>
+                                  )}
                                 </div>
                               </div>
                             </div>
-                          ))}
-                        </div>
-                      </AddSelect.Column>
-                    )}
-                  </AddSelect.Content>
+                          </div>
+                        ))}
+                      </div>
+                    </AddSelect.Column>
+                  )}
                 </AddSelect.Body>
               </Tearsheet.MainContent>
