+++ packages/ibm-products/src/patterns/AddSelect/example/components/SingleAddSelect/SingleAddSelect.tsx
@@ -247,28 +247,26 @@ export const SingleAddSelect = forwardRef<HTMLDivElement, SingleAddSelectProps>(
                   itemCount={filteredItems.length}
                   onSearch={handleSearch}
                 >
-                  <AddSelect.Content>
-                    {filteredItems.length > 0 ? (
-                      filteredItems.map((item) => {
-                        return (
-                          <AddSelect.Row
-                            key={item.id}
-                            itemId={item.id}
-                            title={item.title || ''}
-                            subtitle={item.subtitle}
-                            value={item.value || ''}
-                            icon={item.icon}
-                            disabled={item.disabled}
-                          />
-                        );
-                      })
-                    ) : (
-                      <div className={`${blockClass}__no-results`}>
-                        <h4>{noResultsTitle}</h4>
-                        <p>{noResultsDescription}</p>
-                      </div>
-                    )}
-                  </AddSelect.Content>
+                  {filteredItems.length > 0 ? (
+                    filteredItems.map((item) => {
+                      return (
+                        <AddSelect.Row
+                          key={item.id}
+                          itemId={item.id}
+                          title={item.title || ''}
+                          subtitle={item.subtitle}
+                          value={item.value || ''}
+                          icon={item.icon}
+                          disabled={item.disabled}
+                        />
+                      );
+                    })
+                  ) : (
+                    <div className={`${blockClass}__no-results`}>
+                      <h4>{noResultsTitle}</h4>
+                      <p>{noResultsDescription}</p>
+                    </div>
+                  )}
                 </AddSelect.Body>
               </Tearsheet.MainContent>
             </Tearsheet.Body>
